Passage des stats en float

+++ src/main/java/com/krazytop/leagueoflegends/entity/Runes.java
@@ -2,8 +2,6 @@ package com.krazytop.leagueoflegends.entity;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.databind.JsonNode;
 import lombok.Data;
 
 import java.util.List;
@@ -14,7 +12,7 @@ import java.util.Map;
 public class Runes {
 
     @JsonAlias("statPerks")
-    private Map<String, Integer> stats;
+    private Map<String, Float> stats;
     @JsonAlias("styles")
     private List<RuneCategory> runeCategories;
 }
